requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
PyNaCl==1.5.0
